Add functionality to change status, and edit task
Added the functionality to change the status. The task radio button listens to click event
and toggles the state of any task based on the boolean value of the checkbox. It also toggles the display on
completed task badge

Added form trigger to create an edit form for a task.